Move utility method to new view model rather than the old one
