+++ src/NexusMods.App.UI/Pages/ItemContentsFileTree/New/ViewLoadoutGroupFiles/ViewModel/ViewLoadoutGroupFilesViewModel.cs
@@ -1,4 +1,5 @@
 using System.Reactive.Disposables;
+using DynamicData.Kernel;
 using JetBrains.Annotations;
 using Microsoft.Extensions.Logging;
 using NexusMods.Abstractions.GameLocators;
@@ -11,6 +12,8 @@ using NexusMods.App.UI.Windows;
 using NexusMods.App.UI.WorkspaceSystem;
 using NexusMods.Icons;
 using NexusMods.MnemonicDB.Abstractions;
+using NexusMods.MnemonicDB.Abstractions.ElementComparers;
+using NexusMods.MnemonicDB.Abstractions.IndexSegments;
 using ObservableCollections;
 using R3;
 using ReactiveUI;
@@ -155,5 +158,50 @@ public class ViewLoadoutGroupFilesViewModel : APageViewModel<IViewLoadoutGroupFi
 
             treeDataGridAdapterDisposable.DisposeWith(disposables);
         });
+    }
+    
+    /// <summary>
+    /// Returns the appropriate LoadoutItemGroup of files if the selection contains a LoadoutItemGroup containing files,
+    /// if the selection contains multiple LoadoutItemGroups of files, returns None.
+    /// If the group is completely empty, then it is assumed to be an empty mod and it is returned.
+    /// </summary>
+    internal static Optional<LoadoutItemGroup.ReadOnly> GetViewModFilesLoadoutItemGroup(
+        IReadOnlyCollection<LoadoutItemId> loadoutItemIds, 
+        IConnection connection)
+    {
+        var db = connection.Db;
+        // Only allow when selecting a single item, or an item with a single child
+        if (loadoutItemIds.Count != 1) return Optional<LoadoutItemGroup.ReadOnly>.None;
+        var currentGroupId = loadoutItemIds.First();
+        
+        var groupDatoms = db.Datoms(LoadoutItemGroup.Group, Null.Instance);
+
+        while (true)
+        {
+            var childDatoms = db.Datoms(LoadoutItem.ParentId, currentGroupId);
+            
+            // If no children, assume it's an empty mod and return the group
+            if (childDatoms.Count == 0) return LoadoutItemGroup.Load(db, currentGroupId);
+
+            var childGroups = groupDatoms.MergeByEntityId(childDatoms);
+
+            // We have no child groups, check if children are files
+            if (childGroups.Count == 0)
+            {
+                return LoadoutItemWithTargetPath.TryGet(db, childDatoms[0].E, out _) 
+                    ? LoadoutItemGroup.Load(db, currentGroupId)
+                    : Optional<LoadoutItemGroup.ReadOnly>.None;
+            }
+            
+            // Single child group, check if that group is valid
+            if (childGroups.Count == 1)
+            {
+                currentGroupId = childGroups.First();
+                continue;
+            }
+        
+            // We have multiple child groups, return None
+            if (childGroups.Count > 1) return Optional<LoadoutItemGroup.ReadOnly>.None;
+        }
     }
 }
